finalizing...

+++ src/services/auth/index.js
@@ -1,12 +1,13 @@
 const axios = require('axios');
 import ROUTER from "router";
+const Swal = require('sweetalert2')
 
 export default {
-    user: null,
+    userEmail: null,
     registeredUser: [],
     courses: [],
     setUser(user) {
-        this.user = user
+        this.userEmail = user
     },
 
     register(username, email, password) {
@@ -15,9 +16,12 @@ export default {
             email: email,
             password: password
         }
+        const self = this;
         axios.post('http://localhost:3000/users', data)
             .then(function (response) {
+                localStorage.setItem("user", "dashboard")
                 localStorage.setItem("email", email)
+                self.setUser(email);
                 ROUTER.push('/dashboard')
                 console.log(response);
             })
@@ -32,10 +36,12 @@ export default {
             email: email,
             password: password
         }
+        const self = this;
         axios.post('http://localhost:3000/login', data)
             .then(function (response) {
                 localStorage.setItem("email", email)
-                if (response.data.user !== undefined) {
+                self.setUser(email);
+                if (response.data.user != undefined) {
                     console.log(response.data.user)
                     localStorage.setItem("user", "dashboard")
                     ROUTER.push('/dashboard');
@@ -51,33 +57,38 @@ export default {
             });
     },
     logout() {
-        ROUTER.push('/Homepage')
-        localStorage.clear()
-    },
-    addCourse(course, year, schedule, room, teacher) {
-        this.courses.push({
-            course: course,
-            year: year,
-            schedule: schedule,
-            room: room,
-            teacher: teacher,
+        Swal.fire({
+            title: 'Are you sure you want to logout?',
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes'
+        }).then((result) => {
+            if (result.value) {
+                let timerInterval
+                Swal.fire({
+                    title: 'Logging out ...',
+                    timer: 2000,
+                    timerProgressBar: true,
+                    onBeforeOpen: () => {
+                        Swal.showLoading()
+                    },
+                    onClose: () => {
+                        clearInterval(timerInterval)
+                    }
+                }).then((result) => {
+                    if (
+                        /* Read more about handling dismissals below */
+                        result.dismiss === Swal.DismissReason.timer
+                    ) {
+                        console.log('I was closed by the timer')
+                    }
+                })
+                this.userEmail = null;
+                localStorage.clear()
+                ROUTER.push('/Homepage')
+            }
         });
-        // var p = JSON.parse(JSON.stringify(this.courses))
-        // console.log(p)
-    },
-    save(username, email, password) {
-        for (let i = 0; i < this.registeredUser.length; i++) {
-            this.registeredUser[i].username = username,
-                this.registeredUser[i].email = email,
-                this.registeredUser[i].password = password
-        }
-        // alert('Update Succesfully!')
-        // var p = JSON.parse(JSON.stringify(this.registeredUser))
-        // console.log(p)
-        ROUTER.push('/personalInformation')
-    },
-    // removeCourse(course, year, schedule, room, teacher){
-
-    // }
-
+    }
 }
